// src/types.ts
export type Language = 'fr' | 'ar' | 'en';